+++ src/colors.rs
@@ -196,6 +196,5 @@ static COLORS: ::phf::Map<&'static str, Color> = ::phf::Map {
 };
 
 pub fn rgb_color_from_name(text: &str) -> Option<Color> {
-    let s = text.to_lowercase();
-    COLORS.get(s.as_str()).cloned()
+    COLORS.get(text).cloned()
 }